+++ src/main/java/ro/mxp/food/service/ProductInCartService.java
@@ -4,7 +4,6 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ro.mxp.food.dto.ProductInCartDto;
-import ro.mxp.food.entity.Client;
 import ro.mxp.food.entity.ProductInCart;
 import ro.mxp.food.repository.ClientRepository;
 import ro.mxp.food.repository.ProductInCartRepository;
@@ -12,6 +11,7 @@ import ro.mxp.food.utils.CurrentUsername;
 
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -36,14 +36,14 @@ public class ProductInCartService {
                 .map(cart -> modelMapper.map(cart, ProductInCartDto.class))
                 .collect(Collectors.toList());
 
-        if (clientRepository.findByUsername(currentUsername.displayCurrentUsername()) instanceof Client) {
-            List<ProductInCartDto> productInCartDtoToClient = new LinkedList<>();
+        if (clientRepository.findByUsername(currentUsername.displayCurrentUsername()) != null) {
+            List<ProductInCartDto> productInCartDtoByClient = new LinkedList<>();
             for (ProductInCartDto productInCartDto : productInCartDtoList) {
                 if (productInCartDto.getClient() != null && productInCartDto.getClient().equals(clientRepository.findByUsername(currentUsername.displayCurrentUsername()))) {
-                    productInCartDtoToClient.add(productInCartDto);
+                    productInCartDtoByClient.add(productInCartDto);
                 }
             }
-            return productInCartDtoToClient;
+            return productInCartDtoByClient;
         }
 
         return productInCartDtoList;
@@ -55,11 +55,24 @@ public class ProductInCartService {
     }
 
     public void updateProductInCartService(Long id, int quantity) {
-        productInCartRepository.updateProductInCartRepo(id, quantity);
+        Optional<ProductInCart> productInCart = productInCartRepository.findById(id);
+        ProductInCart thisProductInCart = productInCart.get();
+
+        if ((clientRepository.findByUsername(currentUsername.displayCurrentUsername()) != null)
+                && clientRepository.findByUsername(currentUsername.displayCurrentUsername()).equals(thisProductInCart.getClient())) {
+            productInCartRepository.updateProductInCartRepo(id, quantity);
+        }
+
     }
 
     public void deleteProductInCart(Long id) {
-        productInCartRepository.deleteById(id);
+        Optional<ProductInCart> productInCart = productInCartRepository.findById(id);
+        ProductInCart thisProductInCart = productInCart.get();
+
+        if ((clientRepository.findByUsername(currentUsername.displayCurrentUsername()) != null)
+                && clientRepository.findByUsername(currentUsername.displayCurrentUsername()).equals(thisProductInCart.getClient())) {
+            productInCartRepository.deleteById(id);
+        }
     }
 
 }
